+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Compatibility:
 * Update `File.utime` to return the number of file names in the arguments (#2616, @bjfish).
 * Update `Dir.foreach` to accept an `encoding` parameter (#2627, @bjfish).
 * Update `IO.readlines` to ignore negative limit parameters (#2625 , @bjfish).
+* Update `Math.sqrt` to raise a `Math::DomainError` for negative numbers (#2621, @bjfish).
 
 Performance: